requirements.txt
PyUserInput
mss
Pillow
